+++ packages/http-client-csharp/generator/Microsoft.Generator.CSharp.ClientModel/test/Snippets/TypedSnippetsTests.cs
@@ -1,7 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using System;
 using System.ClientModel;
 using System.Text.Json;
 using Microsoft.Generator.CSharp.ClientModel.Snippets;
@@ -41,28 +40,6 @@ namespace Microsoft.Generator.CSharp.ClientModel.Tests
             Assert.AreEqual(new CSharpType(typeof(JsonSerializer)), result.MethodType);
             Assert.AreEqual(nameof(JsonSerializer.Deserialize), result.MethodName);
 
-        }
-
-        [Test]
-        public void ConvertSnippet_InvokeToDouble()
-        {
-            var arg = Snippet.Literal("2.0");
-            InvokeStaticMethodExpression result = ConvertSnippet.InvokeToDouble(arg);
-
-            Assert.IsNotNull(result.MethodType);
-            Assert.AreEqual(new CSharpType(typeof(Convert)), result.MethodType);
-            Assert.AreEqual(nameof(Convert.ToDouble), result.MethodName);
-        }
-
-        [Test]
-        public void ConvertSnippet_InvokeToInt32()
-        {
-            var arg = Snippet.Literal("2");
-            InvokeStaticMethodExpression result = ConvertSnippet.InvokeToInt32(arg);
-
-            Assert.IsNotNull(result.MethodType);
-            Assert.AreEqual(new CSharpType(typeof(Convert)), result.MethodType);
-            Assert.AreEqual(nameof(Convert.ToInt32), result.MethodName);
         }
 
         [TestCase(false)]
