+++ packages/http-client-csharp/generator/Microsoft.Generator.CSharp/src/TypedSnippets/ConvertSnippet.cs
@@ -2,9 +2,8 @@
 // Licensed under the MIT License.
 using System;
 using Microsoft.Generator.CSharp.Expressions;
-using Microsoft.Generator.CSharp.Snippets;
 
-namespace Microsoft.Generator.CSharp.ClientModel.Snippets
+namespace Microsoft.Generator.CSharp.Snippets
 {
     internal sealed record ConvertSnippet(ValueExpression Untyped) : TypedSnippet(typeof(Convert), Untyped)
     {
