+++ packages/http-client-csharp/generator/Microsoft.Generator.CSharp/test/Snippets/SnippetTests.cs
